// frontend/src/api/model/githubCom4H1RZooraInternalDomainTutorial.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export interface GithubCom4H1RZooraInternalDomainTutorial {
  aparat_hash?: string;
  created_at?: string;
  description_en?: string;
  description_fa?: string;
  id?: string;
  position?: number;
  published_at?: string;
  thumbnail_url?: string;
  title_en?: string;
  title_fa?: string;
  updated_at?: string;
}